Fix API for linking Patient
